# rest/api/v1/api.gen.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error string `json:"error"`
}

// ErroredFile defines model for ErroredFile.
type ErroredFile struct {
	CreatedAt time.Time `json:"created_at"`
	Error     string    `json:"error"`
	Filename  string    `json:"filename"`
	Id        int       `json:"id"`
}

// ErroredFilesResponse defines model for ErroredFilesResponse.
type ErroredFilesResponse struct {
	Files      []ErroredFile `json:"files"`
	Pagination Pagination    `json:"pagination"`
}

// Pagination defines model for Pagination.
type Pagination struct {
	Limit int `json:"limit"`
	Page  int `json:"page"`

	// Total Общее количество записей
	Total int `json:"total"`
}

// Record defines model for Record.
type Record struct {
	Addr      string    `json:"addr"`
	Area      string    `json:"area"`
	Bit       *string   `json:"bit"`
	Block     *string   `json:"block"`
	Class     string    `json:"class"`
	Context   *string   `json:"context"`
	CreatedAt time.Time `json:"created_at"`
	Id        int       `json:"id"`
	InvertBit *string   `json:"invert_bit"`
	Invid     string    `json:"invid"`
	Level     int       `json:"level"`
	Mqtt      *string   `json:"mqtt"`
	MsgId     string    `json:"msg_id"`
	N         int       `json:"n"`
	Text      string    `json:"text"`
	Type      *string   `json:"type"`
	UnitGuid  string    `json:"unit_guid"`
}

// RecordsResponse defines model for RecordsResponse.
type RecordsResponse struct {
	Pagination Pagination `json:"pagination"`
	Records    []Record   `json:"records"`
}

// GetErroredFilesParams defines parameters for GetErroredFiles.
type GetErroredFilesParams struct {
	// Page Номер страницы (начиная с 1)
	Page *int `form:"page,omitempty" json:"page,omitempty"`

	// Limit Количество записей на странице
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`
}

// GetRecordsParams defines parameters for GetRecords.
type GetRecordsParams struct {
	// Page Номер страницы (начиная с 1)
	Page *int `form:"page,omitempty" json:"page,omitempty"`

	// Limit Количество записей на странице
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Получить список файлов с ошибками обработки
	// (GET /api/v1/errors)
	GetErroredFiles(w http.ResponseWriter, r *http.Request, params GetErroredFilesParams)
	// Получить записи по unit_guid с пагинацией
	// (GET /api/v1/records/{unit_guid})
	GetRecords(w http.ResponseWriter, r *http.Request, unitGuid string, params GetRecordsParams)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetErroredFiles operation middleware
func (siw *ServerInterfaceWrapper) GetErroredFiles(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetErroredFilesParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", r.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page", Err: err})
		return
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", r.URL.Query(), &params.Limit)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetErroredFiles(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetRecords operation middleware
func (siw *ServerInterfaceWrapper) GetRecords(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "unit_guid" -------------
	var unitGuid string

	err = runtime.BindStyledParameterWithOptions("simple", "unit_guid", r.PathValue("unit_guid"), &unitGuid, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "unit_guid", Err: err})
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params GetRecordsParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", r.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page", Err: err})
		return
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", r.URL.Query(), &params.Limit)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetRecords(w, r, unitGuid, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of http.ServeMux.
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	ServeHTTP(w http.ResponseWriter, r *http.Request)
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc("GET "+options.BaseURL+"/api/v1/errors", wrapper.GetErroredFiles)
	m.HandleFunc("GET "+options.BaseURL+"/api/v1/records/{unit_guid}", wrapper.GetRecords)

	return m
}